Revert "Refactored server setup and run out of main (#425)"
This reverts commit 05c57c3b8517351d3d434184e3a0697c3bf2f330.
