+++ server/routes_config_loader.go
@@ -49,10 +49,6 @@ func (r *routesConfigLoader) Load(routesConfigFileName string) error {
 }
 
 func (r *routesConfigLoader) Reload() error {
-	if !r.isEnabled() {
-		return nil
-	}
-
 	config, readErr := r.readFile()
 
 	if readErr != nil {
